+++ NEWS.md
@@ -7,7 +7,7 @@ contained within a list structure.
 * Added spant_abfit_benchmark function.
 * Tentative functions for performing "standard" 1H brain analyses: 
 svs_1h_brain_analysis and svs_1h_brain_batch_analysis.
-* Improved support for LCModel on Linux and Windows platforms.
+* Improved support for LCModel analyses.
 
 # spant 1.10.0
 * Fix for NIfTI MRS reader/writer.